requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
ImageHash==4.3.1
magic-filter==1.0.9
multidict==6.0.4
nest-asyncio==1.5.6
numpy==1.24.3
Pillow==9.5.0
psycopg2-binary==2.9.6
pytz==2023.3
PyWavelets==1.4.1
scipy==1.10.1
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
tzlocal==5.0.1
yarl==1.9.2
